Refactor : 함수 나누기 완료 - 페어프로그래밍 JM
