+++ BankManagerConsoleApp/Error/BankError.swift
@@ -15,7 +15,7 @@ extension BankError: LocalizedError {
     var errorDescription: String? {
         switch self {
         case .userInput:
-            return "Invalid Input"
+            return "Error : Invalid Input"
         }
     }
 }
